WIP - starting on session redux to hook conversion
