+++ src/constants.ts
@@ -12,8 +12,6 @@ export const HEARTBEAT_RESPONSE = "HEARTBEAT_RESPONSE";
 
 export const SESSION_AUTHENTICATED = "SESSION_AUTHENTICATED";
 export const SESSION_AUTHENTICATION_FAILED = "SESSION_AUTHENTICATION_FAILED";
-export const SESSION_BUCKETS_REQUEST = "SESSION_BUCKETS_REQUEST";
-export const SESSION_BUCKETS_SUCCESS = "SESSION_BUCKETS_SUCCESS";
 export const SESSION_BUSY = "SESSION_BUSY";
 export const SESSION_COPY_AUTHENTICATION_HEADER =
   "SESSION_COPY_AUTHENTICATION_HEADER";
