+++ src/sagas/index.ts
@@ -15,7 +15,6 @@ export default function* rootSaga(getState: GetStateFn): SagaGen {
     takeEvery(c.SESSION_SETUP, sessionSagas.setupSession, getState),
     takeEvery(c.SESSION_SERVER_CHANGE, sessionSagas.serverChange, getState),
     takeEvery(c.SESSION_GET_SERVERINFO, sessionSagas.getServerInfo, getState),
-    takeEvery(c.SESSION_BUCKETS_REQUEST, sessionSagas.listBuckets, getState),
     takeEvery(c.SESSION_LOGOUT, sessionSagas.sessionLogout, getState),
     takeEvery(
       c.SESSION_COPY_AUTHENTICATION_HEADER,
