+++ src/sagas/session.ts
@@ -8,23 +8,19 @@ import {
 import {
   clearNotifications,
   notifyError,
-  notifyInfo,
   notifySuccess,
 } from "@src/hooks/notifications";
 import { addServer } from "@src/hooks/servers";
 import { DEFAULT_SERVERINFO } from "@src/reducers/session";
-import { clearSession, saveSession } from "@src/store/localStore";
+import { clearSession } from "@src/store/localStore";
 import type {
   ActionType,
   AuthData,
-  BucketEntry,
-  CollectionEntry,
   GetStateFn,
   OpenIDAuth,
   SagaGen,
 } from "@src/types";
-import { clone, copyToClipboard, getAuthLabel } from "@src/utils";
-import { PermissionData } from "kinto/lib/http";
+import { copyToClipboard, getAuthLabel } from "@src/utils";
 import { call, put } from "redux-saga/effects";
 
 export function* serverChange(getState: GetStateFn): SagaGen {
@@ -147,8 +143,6 @@ export function* setupSession(
     yield put(actions.setAuthenticated());
     // Store this valid server url in the history
     addServer(serverInfo.url, fullAuthType);
-
-    yield put(actions.listBuckets());
     yield put(actions.setupComplete(auth));
     notifySuccess("Authenticated.", {
       details: [authLabel],
@@ -180,158 +174,4 @@ export function* sessionCopyAuthenticationHeader(
   const authHeader = getAuthHeader(auth);
   yield call(copyToClipboard, authHeader);
   notifySuccess("Header copied to clipboard");
-}
-
-export function expandBucketsCollections(
-  buckets: BucketEntry[],
-  permissions: PermissionData[]
-): BucketEntry[] {
-  // Create a copy to avoid mutating the source object
-  const bucketsCopy = clone(buckets);
-
-  // Augment the list of bucket and collections with the ones retrieved from
-  // the /permissions endpoint
-  for (const permission of permissions) {
-    if (!Object.prototype.hasOwnProperty.call(permission, "bucket_id")) {
-      // e.g. { resource_name: "root" } permission.
-      continue;
-    }
-    // Add any missing bucket to the current list
-    let bucket = bucketsCopy.find(b => b.id === permission.bucket_id);
-    if (!bucket) {
-      bucket = {
-        id: permission.bucket_id,
-        collections: [],
-        permissions: [],
-        readonly: true,
-        canCreateCollection: true,
-      };
-      bucketsCopy.push(bucket);
-    }
-    // We're dealing with bucket permissions
-    if (permission.resource_name === "bucket") {
-      bucket.permissions = permission.permissions;
-      bucket.readonly = !bucket.permissions.some(bp => {
-        return ["write", "collection:create"].includes(bp);
-      });
-      bucket.canCreateCollection =
-        bucket.permissions.includes("collection:create");
-    }
-    if ("collection_id" in permission) {
-      // Add any missing collection to the current bucket collections list; note
-      // that this will expose collections we have shared records within too.
-      let collection = bucket.collections.find(
-        c => c.id === permission.collection_id
-      );
-      if (!collection) {
-        collection = {
-          id: permission.collection_id,
-          permissions: [],
-          readonly: true,
-        };
-        bucket.collections.push(collection);
-      }
-      // We're dealing with collection permissions
-      if (permission.resource_name === "collection") {
-        collection.permissions = permission.permissions;
-        collection.readonly = !collection.permissions.some(cp => {
-          return ["write", "record:create"].includes(cp);
-        });
-      }
-    }
-  }
-
-  return bucketsCopy;
-}
-
-export function* listBuckets(
-  getState: GetStateFn,
-  action: ActionType<typeof actions.listBuckets>
-): SagaGen {
-  try {
-    const {
-      session: {
-        serverInfo: { user: userInfo, capabilities: serverCapabilities },
-      },
-    } = getState();
-
-    const userBucket = userInfo && userInfo.bucket;
-
-    // Retrieve and build the list of buckets
-    const client = getClient();
-    let data = [];
-    try {
-      data = (yield call([client, client.listBuckets])).data;
-    } catch (error) {
-      // If the user is not allowed to list the buckets, we want
-      // to show an empty list.
-      if (!/HTTP 40[13]/.test(error.message)) {
-        throw error;
-      }
-    }
-
-    // If the default_bucket plugin is enabled, show the Default bucket first in the list.
-    if ("default_bucket" in serverCapabilities && userBucket) {
-      let defaultBucket = data.find(b => b.id == userBucket);
-      if (!defaultBucket) {
-        // It will be shown even if server is empty.
-        defaultBucket = { id: userBucket, last_modified: 0 };
-      }
-      data = [defaultBucket, ...data.filter(b => b.id != userBucket)];
-    }
-
-    const responses = yield call([client, client.batch], batch => {
-      for (const { id } of data) {
-        // When reaching the default bucket by its real id, it does not get created.
-        // https://github.com/Kinto/kinto/issues/1791
-        batch.bucket(id == userBucket ? "default" : id).listCollections();
-      }
-    });
-    let buckets: BucketEntry[] = data.map((bucket, index) => {
-      // Initialize received collections with default permissions and readonly
-      // information.
-      const { data: rawCollections } = responses[index].body;
-      const collections: CollectionEntry[] = rawCollections.map(collection => {
-        return {
-          ...collection,
-          permissions: [],
-          readonly: true,
-        };
-      });
-      // Initialize the list of permissions and readonly flag for this bucket;
-      // when the permissions endpoint is enabled, we'll fill these with the
-      // retrieved data.
-      return {
-        ...bucket,
-        collections,
-        permissions: [],
-        readonly: true,
-        canCreateCollection: true,
-      };
-    });
-
-    // If the Kinto API version allows it, retrieves all permissions
-    if ("permissions_endpoint" in serverCapabilities) {
-      const { data: permissions } = yield call(
-        [client, client.listPermissions],
-        { pages: Infinity, filters: { exclude_resource_name: "record,group" } }
-      );
-      buckets = expandBucketsCollections(buckets, permissions);
-      yield put(actions.permissionsListSuccess(permissions));
-    } else {
-      notifyInfo(
-        [
-          "Permissions endpoint is not enabled on server, ",
-          "listed resources in the sidebar might be incomplete.",
-        ].join("")
-      );
-    }
-
-    yield put(actions.bucketsSuccess(buckets));
-
-    // Save current app state
-    yield call(saveSession, getState().session);
-  } catch (error) {
-    notifyError("Couldn't list buckets.", error);
-  }
 }
