+++ test/sagas/session_test.ts
@@ -5,7 +5,6 @@ import * as notificationHooks from "@src/hooks/notifications";
 import * as serversHooks from "@src/hooks/servers";
 import { DEFAULT_SERVERINFO } from "@src/reducers/session";
 import * as saga from "@src/sagas/session";
-import { saveSession } from "@src/store/localStore";
 import { mockNotifyError, mockNotifySuccess } from "@test/testUtils";
 import { call, put } from "redux-saga/effects";
 
@@ -234,12 +233,6 @@ describe("session sagas", () => {
         );
       });
 
-      it("should retrieve buckets hierarchy", () => {
-        expect(setupSession.next().value).toStrictEqual(
-          put(actions.listBuckets())
-        );
-      });
-
       it("should mark the session setup as completed", () => {
         expect(setupSession.next().value).toStrictEqual(
           put(actions.setupComplete(authData))
@@ -335,191 +328,6 @@ describe("session sagas", () => {
     });
   });
 
-  describe("listBuckets()", () => {
-    let client, listBuckets;
-
-    const serverInfo = {
-      url: "http://server.test/v1",
-      user: {
-        bucket: "defaultBucketId",
-      },
-      capabilities: {
-        permissions_endpoint: {},
-      },
-    };
-
-    const sessionState = { serverInfo };
-
-    beforeAll(() => {
-      client = setClient({
-        batch() {},
-        fetchServerInfo() {},
-        listBuckets() {},
-        listPermissions() {},
-      });
-      const getState = () => ({
-        session: sessionState,
-      });
-      const action = actions.listBuckets();
-      listBuckets = saga.listBuckets(getState, action);
-    });
-
-    describe("Success", () => {
-      it("should fetch the list of buckets", () => {
-        expect(listBuckets.next().value).toStrictEqual(
-          call([client, client.listBuckets])
-        );
-      });
-
-      it("should batch fetch bucket collections list", () => {
-        const buckets = { data: [{ id: "b1" }, { id: "b2" }] };
-
-        expect(listBuckets.next(buckets).value).toHaveProperty(
-          "payload.fn",
-          client.batch
-        );
-      });
-
-      it("should fetch the list of permissions", () => {
-        const responses = [
-          { body: { data: [{ id: "b1c1" }, { id: "b1c2" }] } },
-          { body: { data: [{ id: "b2c1" }] } },
-        ];
-
-        expect(listBuckets.next(responses).value).toStrictEqual(
-          call([client, client.listPermissions], {
-            pages: Infinity,
-            filters: { exclude_resource_name: "record,group" },
-          })
-        );
-      });
-
-      it("should dispatch list of permissions", () => {
-        const permissions = {
-          data: [
-            {
-              bucket_id: "Foo",
-              collection_id: "Bar",
-              resource_name: "bucket",
-              permissions: ["read"],
-            },
-          ],
-        };
-        expect(listBuckets.next(permissions).value).toStrictEqual(
-          put(actions.permissionsListSuccess(permissions.data))
-        );
-      });
-
-      it("should dispatch the list of buckets", () => {
-        expect(listBuckets.next().value).toStrictEqual(
-          put(
-            actions.bucketsSuccess([
-              {
-                id: "b1",
-                collections: [
-                  {
-                    id: "b1c1",
-                    permissions: [],
-                    readonly: true,
-                  },
-                  {
-                    id: "b1c2",
-                    permissions: [],
-                    readonly: true,
-                  },
-                ],
-                permissions: [],
-                readonly: true,
-                canCreateCollection: true,
-              },
-              {
-                id: "b2",
-                collections: [
-                  {
-                    id: "b2c1",
-                    permissions: [],
-                    readonly: true,
-                  },
-                ],
-                permissions: [],
-                readonly: true,
-                canCreateCollection: true,
-              },
-              {
-                id: "Foo",
-                collections: [
-                  {
-                    id: "Bar",
-                    permissions: [],
-                    readonly: true,
-                  },
-                ],
-                permissions: ["read"],
-                readonly: true,
-                canCreateCollection: false,
-              },
-            ])
-          )
-        );
-      });
-
-      it("should save the session", () => {
-        expect(listBuckets.next().value).toStrictEqual(
-          call(saveSession, sessionState)
-        );
-      });
-
-      describe("Forbidden list of buckets", () => {
-        beforeEach(() => {
-          const action = actions.listBuckets();
-          const getState = () => ({
-            session: sessionState,
-          });
-          listBuckets = saga.listBuckets(getState, action);
-
-          listBuckets.next();
-        });
-
-        it("should support 403 errors when fetching list of buckets", () => {
-          // listBucket fails with unauthorized error.
-          listBuckets.throw(new Error("HTTP 403"));
-          // Saga continues without failing.
-          expect(listBuckets.next().value).toHaveProperty(
-            "payload.fn",
-            client.listPermissions
-          );
-        });
-
-        it("should support 401 errors when fetching list of buckets", () => {
-          // listBucket fails with unauthorized error.
-          listBuckets.throw(new Error("HTTP 401"));
-          // Saga continues without failing.
-          expect(listBuckets.next().value).toHaveProperty(
-            "payload.fn",
-            client.listPermissions
-          );
-        });
-      });
-    });
-
-    describe("Failure", () => {
-      it("should dispatch an error notification action", () => {
-        const action = actions.listBuckets();
-        const getState = () => ({
-          session: {
-            serverInfo: { capabilities: [] },
-          },
-        });
-
-        const listBuckets = saga.listBuckets(getState, action);
-        listBuckets.next();
-        const mocked = mockNotifyError();
-        listBuckets.throw("error");
-        expect(mocked).toHaveBeenCalledWith("Couldn't list buckets.", "error");
-      });
-    });
-  });
-
   describe("sessionLogout()", () => {
     let sessionLogout;
 
@@ -540,71 +348,5 @@ describe("session sagas", () => {
     it("should reset the client", () => {
       expect(() => getClient()).toThrow(Error, /not configured/);
     });
-  });
-});
-
-describe("expandBucketsCollections()", () => {
-  const buckets = [
-    { id: "b1", permissions: [], collections: [], readonly: undefined },
-    { id: "b2", permissions: [], collections: [], readonly: undefined },
-  ];
-
-  function bucketPerm(bucket_id, permissions) {
-    return { resource_name: "bucket", bucket_id, permissions };
-  }
-
-  function collectionPerm(bucket_id, collection_id, permissions) {
-    return {
-      resource_name: "collection",
-      bucket_id,
-      collection_id,
-      permissions,
-    };
-  }
-
-  const permissions = [
-    { resource_name: "root", permissions: ["bucket:create"] },
-    bucketPerm("b1", ["read"]),
-    bucketPerm("b2", ["read"]),
-    bucketPerm("b4", ["write"]),
-    collectionPerm("b1", "b1c1", ["read", "write"]),
-    collectionPerm("b2", "b2c1", ["read"]),
-    collectionPerm("b3", "b3c1", ["read", "write"]),
-    bucketPerm("foo", ["read"]),
-    collectionPerm("foo", "foo", ["read"]),
-  ];
-
-  const tree = saga.expandBucketsCollections(buckets, permissions, 2);
-
-  it("should denote a bucket as writable", () => {
-    expect(tree.find(b => b.id === "b4").readonly).toBe(false);
-  });
-
-  it("should denote a bucket as readonly", () => {
-    expect(tree.find(b => b.id === "b2").readonly).toBe(true);
-  });
-
-  it("should denote a collection as writable", () => {
-    const b1c1 = tree
-      .find(b => b.id === "b1")
-      .collections.find(c => c.id === "b1c1");
-    expect(b1c1.readonly).toBe(false);
-  });
-
-  it("should denote a collection as readonly", () => {
-    const b2c1 = tree
-      .find(b => b.id === "b2")
-      .collections.find(c => c.id === "b2c1");
-    expect(b2c1.readonly).toBe(true);
-  });
-
-  it("should infer an implicit bucket", () => {
-    expect(tree.find(b => b.id === "b3").readonly).toBe(true);
-  });
-
-  it("should distinguish resource ids", () => {
-    const fooBucket = tree.find(b => b.id === "foo");
-    expect(fooBucket).toBeDefined();
-    expect(fooBucket.collections.find(c => c.id === "foo")).toBeDefined();
   });
 });
